requirements.txt
notion-client
python-dotenv         
pydantic             
requests 
streamlit
streamlit-option-menu
streamlit-tags
plotly
pandas
google-generativeai
ollama
rich
loguru
typer

#Development
pytest
black